Refactor: structure Frame
