+++ src/components/Mypage/Mypage.jsx
@@ -2,13 +2,13 @@ import React, { useEffect, useState } from "react";
 import { animated, useSpring } from "react-spring";
 import { useHistory } from "react-router";
 import PropTypes from "prop-types";
-import Title from "@frames/Title/Title";
-import WhiteContainer from "@frames/WhiteContainer/WhiteContainer";
+import Title from "components/common/Title";
+import WhiteContainer from "components/common/WhiteContainer";
 import PopupSparcs from "./PopupSparcs/PopupSparcs";
 import PopupPolicy from "./PopupPolicy/PopupPolicy";
 import PopupMypage from "./PopupMypage/PopupMypage";
 import ProfileImg from "./ProfileImg";
-import useTaxiAPI from "@components/Frame/useTaxiAPI/useTaxiAPI";
+import useTaxiAPI from "hooks/useTaxiAPI";
 import axios from "@tools/axios";
 
 import AssignmentRoundedIcon from "@material-ui/icons/AssignmentRounded";
