+++ src/components/Search/Search/Search.jsx
@@ -2,8 +2,8 @@ import React, { useState, useRef, useEffect } from "react";
 import { animated, useSpring } from "react-spring";
 import { useR2state } from "hooks/useReactiveState";
 import RLayout from "components/common/RLayout";
-import Title from "@frames/Title/Title";
-import SubmitButton from "@frames/SubmitButton/SubmitButton";
+import Title from "components/common/Title";
+import SubmitButton from "components/common/roomOptions/SubmitButton";
 import SideResult from "../SearchResult/SideResult";
 import axios from "@tools/axios";
 import moment from "moment";
