+++ src/components/Search/SearchResult/SideResult.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import WhiteContainer from "@frames/WhiteContainer/WhiteContainer";
+import WhiteContainer from "components/common/WhiteContainer";
 import Room from "@components/Room/Room/RoomElement";
 import RoomList from "@components/Room/Room/RoomList";
 import PropTypes from "prop-types";
