+++ src/components/common/RLayout.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useRef, useEffect } from "react";
+import React from "react";
 import { useR1state, useR2state, usePopupstate } from "hooks/useReactiveState";
 import PropTypes from "prop-types";
